feat: made player script compute hook a bit more ergonomic to use DX wise, this also makes it feel more inline with what the industry does for JS / TS workers
We are abusing the fact that when `deno_emit` is used to bundle a script in IIFE ("classic") mode.

It wraps the bundled content in a function that returns the exports of the entry point as a plain JavaScript object. And then it calls the function immediately.

The `node:vm` compatbility module's `Script.runInContext` returns that last statement / value of the script executed. Meaning we always get the value of the returned exports.